Allow sale of multiple LAND together
Here are some changes I made to allow the sale of multiple parcels together. I apologize for so many changes at once, I have never worked in a team before.

It seemed to me that the auctionId property is not used ever, even indexed for later. The hash is created at a time that is not useful to check data integrity, and the ID is only output in the events. I cannot see a purpose for this, but added a comment at the very end of how it would fit with what I have done, should it be necessary for reasons unknown to me.

I left the Auction as the only struct, as adding as Estate struct would require a separate mapping where lands and orders would be divided between the two, so this did not seem ideal. Instead I changed assetId in the struct to assets[], so that it could be used for single or multi-asset auctions. To accommodate this, in line 125 I initialized a single unit array to pass into the created auction.

In the events, took out auction Id, added bool indexed isEstate. Every event that happens on a single land sale will be categorized as such. In an estate sale, a separate event will trigger for each parcel, but the event will contain estate = true to show that it was part of an estate event. In addition, an estate event triggers that contains the complete array of assetIds involved.

Added requirement that assetId must not be mapped to Auction before an order can be created. This was okay with single asset orders, but must be prevented with multi-asset. Line 120, 167.

createEstate() is passed an array of Ids, stores a couple local items to prevent excessive calls in a for loop.I believe I worked it so that for loops create no extra gas use than for each land to be executed separately, in fact I believe it is well optimized above that as not all require statements need be checked on every loop. The logic should follow that multiple addresses cannot make a single call involving a group of assets, so checking through the assets once shows that those assets are qualified with the valid address and, in later functions, only the address of one of them needs to be checked.

In executeOrder() and executeEstate() reads to the mapping properties of seller and price were called many times throughout, so I placed a local memory item where I believe it would be most optimal for all but outlying cases. There were, however a few lines that I do not know if they are necessary. I left them just in case but I will bring them to your attention in case it can reduce gas:
- 244 and 294 - Why would these conditions be needed? Even if burned,  how would the land be posted for sale from 0x0?
- 245 and 295 - Why can the address not purchase their own land? They will simply lose money and give commission to the developers!
- combinations of 249/251 and 299/301 - When the Auction is created, it is checked that the seller recorded is the owner of the assets, therefore the same check here seems redundant and just costs extra gas.

This was related to my own desires for the market and related to issue#84 "few parcel in one publication"